third_party/WebKit/LayoutTests/inspector/timeline/timeline-frame-model.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>
function test()
{
    var testData = {
        "main thread only": [
            {"startTime": 1000.000, "data":{},"endTime":1003.999, "type":"Program", "children":[
                {"startTime":1000.001, "type":"BeginFrame", "data":{"id":"1"}},
                {"startTime":1000.002, "endTime":1003.999, "type":"FunctionCall" }
            ]},
            {"startTime": 1016.000, "data":{},"endTime":1026.999, "type":"Program", "children":[
                {"startTime":1016.001, "type":"BeginFrame", "data":{"id":"2"}},
                {"startTime":1016.002, "endTime":1018.999, "type":"FunctionCall", "data":{}},
                {"startTime":1019.000, "endTime":1020.999, "type":"StyleRecalculate", "data":{}},
                {"startTime":1021.000, "endTime":1026.999, "type":"Layout", "data":{}},
            ]},
            {"startTime": 1032.000, "data":{},"endTime":1033.999, "type":"Program", "children":[
                {"startTime":1032.001, "type":"BeginFrame", "data":{"id":"3"}},
            ]},
            {"startTime": 1048.000, "data":{},"endTime":1049.999, "type":"Program", "children":[
                {"startTime":1048.001, "type":"BeginFrame", "data":{"id":"4"}},
            ]}
        ],
        "impl-side only": [
            {"startTime":1000.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1016.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1030.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1032.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1046.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1048.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1064.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1065.001, "endTime":1066, "type":"Paint", "thread": "paintWorker", "data":{}},
            {"startTime":1066.001, "endTime":1067, "type":"Paint", "thread": "paintWorker", "data":{}},
            {"startTime":1067.001, "endTime":1068, "type":"Paint", "thread": "paintWorker", "data":{}},
            {"startTime":1078.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1080.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1081.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1082.000, "data":{},"endTime":1082.1, "type":"Program", "children":[]}
        ],
        "impl-side with commits": [
            {"startTime":1000.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1000.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1014.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1016.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1030.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1032.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime": 1001.000, "data":{},"endTime":1033.999, "type":"Program", "children":[
                {"startTime":1001.001, "type":"BeginFrame", "data":{"id":"1"}},
                {"startTime":1001.002, "endTime":1018.999, "type":"FunctionCall", "data":{}},
                {"startTime":1019.000, "endTime":1020.999, "type":"StyleRecalculate", "data":{}},
                {"startTime":1021.000, "endTime":1032.999, "type":"Layout", "data":{}},
                {"startTime":1032.000, "endTime":1033.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime": 1034.000, "data":{},"endTime":1039.999, "type":"Program", "children":[
                {"startTime":1034.002, "endTime":1039.999, "type":"FunctionCall", "data":{}},
            ]},
            {"startTime":1045.001, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"1"}},
            {"startTime":1046.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1048.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1048.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1060.001, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"2"}},
            {"startTime":1062.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1049.000, "data":{},"endTime":1060.999, "type":"Program", "children":[
                {"startTime":1050.001, "type":"BeginFrame", "data":{"id":"2"}},
                {"startTime":1050.002, "endTime":1060.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime":1064.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1064.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime": 1065.000, "data":{},"endTime":1078.999, "type":"Program", "children":[
                {"startTime":1065.001, "type":"BeginFrame"}
            ]}
        ],
        "impl-side with interleaving commits": [
            {"startTime":1000.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1000.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1014.000, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"1"}},
            {"startTime":1014.001, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1001.000, "data":{},"endTime":1013.999, "type":"Program", "children":[
                {"startTime":1001.001, "type":"BeginFrame", "data":{"id":"1"}},
                {"startTime":1001.002, "endTime":1013.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime":1016.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1016.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1030.000, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"2"}},
            {"startTime":1030.001, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1017.000, "data":{},"endTime":1029.999, "type":"Program", "children":[
                {"startTime":1017.001, "type":"BeginFrame", "data":{"id":"2"}},
                {"startTime":1017.002, "endTime":1025.999, "type":"FunctionCall", "data":{}},
                {"startTime":1026.000, "endTime":1029.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime":1032.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1032.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1046.000, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"3"}},
            {"startTime":1046.001, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1033.000, "data":{},"endTime":1045.999, "type":"Program", "children":[
                {"startTime":1033.001, "type":"BeginFrame", "data":{"id":"3"}},
                {"startTime":1033.002, "endTime":1040.999, "type":"CompositeLayers", "data":{}},
                {"startTime":1041.002, "endTime":1045.999, "type":"FunctionCall", "data":{}},
            ]},
            {"startTime":1048.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime": 1049.000, "data":{},"endTime":1049.999, "type":"Program", "children":[
            ]}
        ],
        "pre-frame time accounting": [
            {"startTime": 1000.000, "data":{},"endTime":1029.999, "type":"Program", "children":[
                {"startTime":1000.002, "endTime":1028.999, "type":"FunctionCall", "data":{}},
                {"startTime":1029.000, "endTime":1029.999, "type":"ScheduleStyleRecalculation", "data":{}},
            ]},
            {"startTime":1032.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1032.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1034.001, "type":"ActivateLayerTree", "thread": "implSide", "data":{"id": "1"}},
            {"startTime":1035.001, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1033.000, "data":{},"endTime":1033.999, "type":"Program", "children":[
                {"startTime":1033.001, "type":"BeginFrame", "data":{"id":"1"}},
                {"startTime":1033.002, "endTime":1033.400, "type":"RecalculateStyles", "data":{}},
                {"startTime":1033.401, "endTime":1033.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime":1048.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime": 1050.000, "data":{},"endTime":1064.999, "type":"Program", "children":[
                {"startTime":1050.002, "endTime":1058.999, "type":"FunctionCall", "data":{}},
                {"startTime":1059.000, "type":"InvalidateLayout", "data":{}},
            ]},
            {"startTime":1063.001, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1064.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1064.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1071.001, "type":"ActivateLayerTree", "thread": "implSide", "data":{"id": "2"}},
            {"startTime":1071.002, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1065.000, "data":{},"endTime":1070.999, "type":"Program", "children":[
                {"startTime":1065.001, "type":"BeginFrame", "data":{"id":"2"}},
                {"startTime":1065.002, "endTime":1068.000, "type":"Layout", "data":{}},
                {"startTime":1068.001, "endTime":1070.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime":1080.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1080.001, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1081.001, "type":"ActivateLayerTree", "thread": "implSide", "data":{"id": "3"}},
            {"startTime":1081.002, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1073.000, "data":{},"endTime":1080.999, "type":"Program", "children":[
                {"startTime":1073.002, "endTime":1079.999, "type":"ScrollLayer", "data":{}},
            ]},
            {"startTime":1096.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1096.001, "type":"ActivateLayerTree", "thread": "implSide", "data":{"id": "3"}},
            {"startTime":1096.002, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1081.000, "data":{},"endTime":1081.999, "type":"Program", "children":[
                {"startTime":1081.001, "type":"BeginFrame", "data":{"id":"3"}},
                {"startTime":1081.001, "endTime":1081.999, "type":"CompositeLayers", "data":{}},
            ]}
        ],
        "record processing order": [
            {"startTime":1000.000, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1000.001, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1000.002, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1013.003, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"1"}},
            {"startTime":1013.005, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1016.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime": 1001.000, "data":{},"endTime":1016.999, "type":"Program", "children":[
                {"startTime":1001.001, "type":"BeginFrame", "data":{"id":"1"}},
                {"startTime":1001.002, "endTime":1012.999, "type":"FunctionCall", "data":{}},
                {"startTime":1013.002, "endTime":1016.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime":1030.000, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"1"}},
            {"startTime":1030.001, "type":"DrawFrame", "thread": "implSide"},
            {"startTime":1031.000, "type":"BeginFrame", "thread": "implSide"},
            {"startTime":1031.002, "type":"RequestMainThreadFrame", "thread": "implSide"},
            {"startTime":1061.000, "type":"ActivateLayerTree", "thread": "implSide","data":{"id":"1"}},
            {"startTime":1062.001, "type":"DrawFrame", "thread": "implSide"},
            {"startTime": 1032.000, "data":{},"endTime":1063.999, "type":"Program", "children":[
                {"startTime":1032.001, "type":"BeginFrame", "data":{"id":"1"}},
                {"startTime":1032.002, "endTime":1042.999, "type":"FunctionCall", "data":{}},
                {"startTime":1043.002, "endTime":1062.999, "type":"CompositeLayers", "data":{}},
            ]},
            {"startTime":1080.001, "type":"DrawFrame", "thread": "implSide"}
        ]
    };

    var timelinePanel = WebInspector.inspectorView._showPanel("timeline");
    timelinePanel._overviewModeSetting.set("Frames");
    timelinePanel._onModeChanged();
    var overview = timelinePanel._overviewPane;
    var frameModel = overview._overviewControls[0]._frameModel;

    for (var testName in testData) {
        var data = testData[testName];
        InspectorTest.loadTimelineRecords(data);
        InspectorTest.addResult("Test: " + testName);
        frameModel.frames().forEach(InspectorTest.dumpFrame, InspectorTest);
    }
    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
Test timeline records coalescing
</p>

</body>
</html>
