third_party/WebKit/LayoutTests/inspector/timeline/timeline-frame-model-expected.txt

Test timeline records coalescing

Test: main thread only
{
    cpuTime : 3.999
    duration : 16
    endTime : 1016.001
    startTime : 1000.001
    timeByCategory : {
        other : 0.002
        rendering : 0
        scripting : 3.997
    }
}
{
    cpuTime : 10.999
    duration : 16
    endTime : 1032.001
    startTime : 1016.001
    timeByCategory : {
        other : 2.003
        rendering : 5.999
        scripting : 2.997
    }
}
{
    cpuTime : 1.999
    duration : 16
    endTime : 1048.001
    startTime : 1032.001
    timeByCategory : {
        other : 1.999
        rendering : 0
    }
}
Test: impl-side only
{
    cpuTime : 0
    duration : 30
    endTime : 1030
    startTime : 1000
    timeByCategory : {
        rendering : 0
    }
}
{
    cpuTime : 0
    duration : 16
    endTime : 1046
    startTime : 1030
    timeByCategory : {
        other : 0
        rendering : 0
    }
}
{
    cpuTime : 2.997
    duration : 32
    endTime : 1078
    startTime : 1046
    timeByCategory : {
        other : 0
        painting : 2.997
        rendering : 0
    }
}
{
    cpuTime : 0
    duration : 3
    endTime : 1081
    startTime : 1078
    timeByCategory : {
        other : 0
        rendering : 0
    }
}
Test: impl-side with commits
{
    cpuTime : 32.999
    duration : 46
    endTime : 1046
    startTime : 1000
    timeByCategory : {
        other : 1.004
        painting : 1.999
        rendering : 11.999
        scripting : 17.997
    }
}
{
    cpuTime : 11.999
    duration : 16
    endTime : 1062
    startTime : 1046
    timeByCategory : {
        other : 1.002
        painting : 10.997
        rendering : 0
    }
}
Test: impl-side with interleaving commits
{
    cpuTime : 12.999
    duration : 14.001
    endTime : 1014.001
    startTime : 1000
    timeByCategory : {
        other : 0.002
        painting : 12.997
        rendering : 0
    }
}
{
    cpuTime : 12.999
    duration : 16
    endTime : 1030.001
    startTime : 1014.001
    timeByCategory : {
        other : 0.003
        painting : 3.999
        rendering : 0
        scripting : 8.997
    }
}
{
    cpuTime : 8.002
    duration : 16
    endTime : 1046.001
    startTime : 1030.001
    timeByCategory : {
        other : 0.005
        painting : 7.997
        rendering : 0
    }
}
Test: pre-frame time accounting
{
    cpuTime : 0.999
    duration : 3.001
    endTime : 1035.001
    startTime : 1032
    timeByCategory : {
        other : 0.003
        painting : 0.598
        rendering : 0.398
    }
}
{
    cpuTime : 0
    duration : 28
    endTime : 1063.001
    startTime : 1035.001
    timeByCategory : {
        other : 0
        rendering : 0
    }
}
{
    cpuTime : 20.998
    duration : 8.001
    endTime : 1071.002
    startTime : 1063.001
    timeByCategory : {
        other : 6.005
        painting : 2.998
        rendering : 2.998
        scripting : 8.997
    }
}
Test: record processing order
{
    cpuTime : 15.999
    duration : 13.005
    endTime : 1013.005
    startTime : 1000
    timeByCategory : {
        other : 0.005
        painting : 3.997
        rendering : 0
        scripting : 11.997
    }
}
{
    cpuTime : 0
    duration : 16.996
    endTime : 1030.001
    startTime : 1013.005
    timeByCategory : {
        other : 0
        rendering : 0
    }
}
{
    cpuTime : 31.999
    duration : 32
    endTime : 1062.001
    startTime : 1030.001
    timeByCategory : {
        other : 1.005
        painting : 19.997
        rendering : 0
        scripting : 10.997
    }
}
